+++ test/e2e/modules/resources/rd/pod_group/pod_group.go
@@ -83,7 +83,7 @@ func IsNotReadyForScheduling(event *v1.Event) bool {
 		return false
 	}
 	match, err := regexp.MatchString(
-		"Job is not ready for scheduling. Waiting for \\d pods, currently \\d existing",
+		"Job is not ready for scheduling. Waiting for \\d pods, currently \\d exist, \\d are gated",
 		event.Message)
 	Expect(err).To(Succeed())
 	return match